Mob & Zone Autoscaling (#132)
# Motivation

The goal for this is to support "Auto Scaling" for mobs in areas.

Many MUD's, including the "default" GoMud, target a difficulty curve that radiates outward from the central game-start zone. To support this, it's typical to make mobs more difficult by adjusting type, skills, equipment, etc... but also the overall "level" of the mob, which has the most far reaching consequences of difficult.

This can be difficult to balance as one builds out areas, so as an *optional* feature, a level override can be defined on a per-area basis.

This can also potentially serve as a data point for area difficulty.

It's important to be able to override the zone auto scaling ranges, so that mobs can be created that are particularly easy or difficult. 

These changes are being pushed through for purposes of deployment and testing in a live environment.

# Changes
* `zone` admin command for viewing/changing zone settings
* Created new zones - `Frostfang Slums` and `Stormwatchers Keep` - formerly part of other zones, now partitioned for better organization and customization
* Customized some of the zones with level ranges. This remains to be tested for balance.
* Organized some structs into own files in the rooms package.