[Update] docker build process (#197)
Volte6,

This wasn't asked for or probably necessary for your normal development, so feel free to close/remove/request changes if you like. 

This PR updates the docker compose to a simple command which will run all of the services. It removes some of the bloat of `entrypoint.sh` files and compile steps and doesn't use any of the `make` targets (since it didn't make sense to compile for arm for instance on a docker base image). 

I removed `gofmt` requirement that was present when compiling the `go-mud-server` for building because `gofmt -l .` returns many incorrectly formatted files. This may have been configuration/error on my part since I'm not very familiar with go.

Anyway, I hope this is helpful and if it's not that's no problem at all.

Best,
Colin